Library improvements

+++ src/tests/DFR0592/dfr0592.h
@@ -158,7 +158,7 @@ int encoder_get_speed(const struct dfr_board *board, int motor, int *speed);
  *
  * In the event of failure, check errno to get the cause.
  */
-int motor_set_speed(const struct dfr_board *board, int motor, int speed);
+int motor_set_speed(const struct dfr_board *board, int motor, float speed);
 
 /*
  * Stop the motor movement
